+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 ### Added
 
+- ✨ Synchronize support resources ([#22]) ([**@denialhaag**])
 - ✨ Synchronize issue templates ([#21]) ([**@denialhaag**])
 
 ## [1.0.0]
@@ -30,6 +31,7 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 <!-- PR links -->
 
+[#22]: https://github.com/munich-quantum-toolkit/templates/pull/21
 [#21]: https://github.com/munich-quantum-toolkit/templates/pull/21
 [#13]: https://github.com/munich-quantum-toolkit/templates/pull/13
 [#1]: https://github.com/munich-quantum-toolkit/templates/pull/1